+++ next.config.ts
@@ -1,4 +1,5 @@
 import type { NextConfig } from "next";
+import withSerwist from "@serwist/next";
 import { version } from "./package.json";
 
 const nextConfig: NextConfig = {
@@ -43,4 +44,8 @@ const nextConfig: NextConfig = {
   },
 };
 
-export default nextConfig;
+export default withSerwist({
+  swSrc: "app/sw.ts",
+  swDest: "public/sw.js",
+  disable: process.env.NODE_ENV === "development",
+})(nextConfig);